+++ lib/__init__.py
@@ -51,6 +51,9 @@ from . import reg
 from . import dat
 # import the units module
 from . import units
+# By default, do not import the module for handling special applications
+# This module has requirements beyond the base pyromat installation
+#from . import aps
 
 reg.regload()
 dat.load()